egui_demo_app: move web-part to own module
